Feat: Navbar 데이터 삽입
