+++ packages/web/src/components/Navbar/Navbar.module.css
@@ -1,40 +1,29 @@
 .nav {
-  height: 80px;
+  height: 100px;
   width: 100%;
-  margin: 10px 0;
-  padding: 0 80px;
+  padding: 20px 54px;
+
   display: flex;
   flex-direction: row;
   align-items: center;
+
+  border-bottom: 1px solid #ddd;
+  background: var(--White-50, #fff);
 }
 
 .nav_logo {
-  width: 200px;
+  width: 120px;
   height: 100%;
-  border: solid black;
-}
-
-.nav_item {
   display: flex;
-  flex-direction: row;
-
-  & li {
-    width: 100px;
-    border: solid red;
-    margin-left: 16px;
-    text-align: center;
-
-    &:first-child {
-      margin-left: 20px;
-    }
-  }
+  flex-direction: column;
+  align-content: flex-start;
+  justify-content: center;
 }
 
 .nav_mode {
   width: 48px;
   height: 48px;
   margin-right: 16px;
-  background-color: red;
 }
 
 .nav_login {
